requirements.txt
# ============================================
# UI Framework
# ============================================
streamlit==1.41.1

# ============================================
# Gemini API
# ============================================
google-generativeai==0.8.3

# ============================================
# Document Generation
# ============================================
python-docx==1.1.2
Jinja2==3.1.4
Pillow==11.0.0

# ============================================
# Diagram Rendering & PDF Reading
# ============================================
requests==2.32.3
pypdf==5.1.0

# ============================================
# Data Validation
# ============================================
pydantic==2.10.3

# ============================================
# Environment & Config
# ============================================
python-dotenv==1.0.1
